# Much faster navigator index creation for mixed Swift and Objective-C projects