… Sources/SwiftDocC/LinkTargets/LinkDestinationSummary.swift
… Sources/SwiftDocC/Model/Rendering/References/TopicImage.swift
… Sources/SwiftDocC/Model/Rendering/Variants/VariantPatchOperation.swift
… Tests/SwiftDocCTests/Indexing/NavigatorIndexTests.swift
… Tests/SwiftDocCTests/Model/RenderNodeSerializationTests.swift
… Tests/SwiftDocCTests/Model/SemaToRenderNodeTests.swift
… Tests/SwiftDocCTests/Rendering/RenderNodeTranslatorTests.swift
… Tests/SwiftDocCTests/Rendering/Variants/VariantPatchOperationTests.swift
… Tests/SwiftDocCUtilitiesTests/OutOfProcessReferenceResolverTests.swift