<!--
If you're opening a PR to cherry-pick a change for a release branch, use this template instead:
https://github.com/apple/swift-docc/blob/main/.github/PULL_REQUEST_TEMPLATE/CHERRY_PICK.md
-->

Bug/issue #, if applicable: rdar://127759734

## Summary

This fixes a major inefficiency in the navigator index creation for large (10k+ symbols) mixed Swift and Objective-C projects. 

With these changes impacted projects will build about 5 times faster total `docc convert` time. Even larger projects may see even faster builds (the biggest improvement that I've seen was ~30 times faster in one project).

### Benchmarks

Below are two benchmarks from two large mixed Swift and Objective-C projects that show a 6 times improvement and a 5 time improvement compared to current main:

```
┌──────────────────────────────────────────────────────────────────────────────────────────────────────────┐
│ Metric                                   │ Change          │ main                 │ current              │
├──────────────────────────────────────────────────────────────────────────────────────────────────────────┤
│ Duration for 'bundle-registration'       │ no change¹      │ 8.353 sec            │ 8.271 sec            │
│ Duration for 'convert-total-time'        │ -82.9%²         │ 89.014 sec           │ 15.223 sec           │
│ Duration for 'documentation-processing'  │ -93.41%³        │ 78.942 sec           │ 5.202 sec            │
│ Duration for 'finalize-navigation-index' │ no change⁴      │ 1.031 sec            │ 1.036 sec            │
│ Peak memory footprint                    │ no change⁵      │ 1.03 GB              │ 1.02 GB              │
│ Data subdirectory size                   │ no change⁶      │ 399.7 MB             │ 400.4 MB             │
│ Index subdirectory size                  │ no change⁷      │ 24.5 MB              │ 24.4 MB              │
│ Total DocC archive size                  │ no change⁸      │ 456.8 MB             │ 458.5 MB             │
│ Topic Anchor Checksum                    │ no change       │ 7da1d0d3ff251d8c2c0f │ 7da1d0d3ff251d8c2c0f │
│ Topic Graph Checksum                     │ change          │ 80f03eee58720209c93f │ 542c6795a3c9d1b555c3 │
└──────────────────────────────────────────────────────────────────────────────────────────────────────────┘
```

```
┌──────────────────────────────────────────────────────────────────────────────────────────────────────────┐
│ Metric                                   │ Change          │ main                 │ current              │
├──────────────────────────────────────────────────────────────────────────────────────────────────────────┤
│ Duration for 'bundle-registration'       │ no change¹,²    │ 3.052 sec            │ 2.954 sec            │
│ Duration for 'convert-total-time'        │ -80.11%³,⁴      │ 35.358 sec           │ 7.034 sec            │
│ Duration for 'documentation-processing'  │ -88.23%⁵,⁶      │ 31.986 sec           │ 3.765 sec            │
│ Duration for 'finalize-navigation-index' │ no change⁷,⁸    │ 0.169 sec            │ 0.163 sec            │
│ Peak memory footprint                    │ no change⁹      │ 577.3 MB             │ 575.6 MB             │
│ Data subdirectory size                   │ no change¹⁰     │ 108.6 MB             │ 108.6 MB             │
│ Index subdirectory size                  │ no change¹¹     │ 3.9 MB               │ 3.9 MB               │
│ Total DocC archive size                  │ no change¹²     │ 138.6 MB             │ 138.6 MB             │
│ Topic Anchor Checksum                    │ no change       │ d41d8cd98f00b204e980 │ d41d8cd98f00b204e980 │
│ Topic Graph Checksum                     │ change          │ d4e127bf2e5d126c1cb7 │ 3d1b0f2fcacafee063b6 │
└──────────────────────────────────────────────────────────────────────────────────────────────────────────┘
```

If you look at these closely you may see that the output size and topic graph checksum aren't the same. At first I thought that I caused a regression but it turns out that the output isn't fully stable for mixed Swift and Objective-C projects. I tested this on "main", "release/6.0", and "release/5.10" and they all show similar differences when comparing the same commit of DocC to itself. For example, these 3 measurements are all gathered from the current main (`664acfd7`) and show similar changes in output size and topic graph checksum:
```
┌──────────────────────────────────────────────────────────────────────────────────────────────────────────┐
│ Metric                                   │ Change          │ main-1               │ main-2               │
├──────────────────────────────────────────────────────────────────────────────────────────────────────────┤
│ Duration for 'bundle-registration'       │ no change¹,²    │ 8.242 sec            │ 8.41 sec             │
│ Duration for 'convert-total-time'        │ no change³,⁴    │ 96.103 sec           │ 96.014 sec           │
│ Duration for 'documentation-processing'  │ no change⁵,⁶    │ 84.831 sec           │ 84.838 sec           │
│ Duration for 'finalize-navigation-index' │ no change⁷,⁸    │ 1.104 sec            │ 1.09 sec             │
│ Peak memory footprint                    │ no change⁹      │ 1.03 GB              │ 1.05 GB              │
│ Data subdirectory size                   │ no change¹⁰     │ 403.9 MB             │ 399.9 MB             │
│ Index subdirectory size                  │ no change¹¹     │ 24.5 MB              │ 24.4 MB              │
│ Total DocC archive size                  │ no change¹²     │ 462 MB               │ 458 MB               │
│ Topic Anchor Checksum                    │ no change       │ 15bea17f1e7688a3fc2d │ 15bea17f1e7688a3fc2d │
│ Topic Graph Checksum                     │ change          │ 5b0bf590353b64a56db7 │ 477075a5259bee77ca9c │
└──────────────────────────────────────────────────────────────────────────────────────────────────────────┘
```

```
┌──────────────────────────────────────────────────────────────────────────────────────────────────────────┐
│ Metric                                   │ Change          │ main-2               │ main-3               │
├──────────────────────────────────────────────────────────────────────────────────────────────────────────┤
│ Duration for 'bundle-registration'       │ no change¹,²    │ 8.41 sec             │ 8.324 sec            │
│ Duration for 'convert-total-time'        │ no change³,⁴    │ 96.014 sec           │ 96.647 sec           │
│ Duration for 'documentation-processing'  │ no change⁵,⁶    │ 84.838 sec           │ 85.55 sec            │
│ Duration for 'finalize-navigation-index' │ no change⁷      │ 1.09 sec             │ 1.101 sec            │
│ Peak memory footprint                    │ no change⁸,⁹    │ 1.05 GB              │ 1.05 GB              │
│ Data subdirectory size                   │ no change¹⁰     │ 399.9 MB             │ 409 MB               │
│ Index subdirectory size                  │ no change¹¹     │ 24.4 MB              │ 24.4 MB              │
│ Total DocC archive size                  │ no change¹²     │ 458 MB               │ 467.1 MB             │
│ Topic Anchor Checksum                    │ no change       │ 15bea17f1e7688a3fc2d │ 15bea17f1e7688a3fc2d │
│ Topic Graph Checksum                     │ change          │ 477075a5259bee77ca9c │ eb532ce04afad3fdfcfd │
└──────────────────────────────────────────────────────────────────────────────────────────────────────────┘
```

I've gone over my changes many times and since many previous releases have the same behavior, I'm reasonably certain that I didn't cause them now.

### Why this is faster

The added comment in NavigatorIndex.swift explain the specific changes in more detail, but long story short:

Using `RenderNodeVariantOverridesApplier()` to get the Objective-C variant of a `RenderNode` is very slow and involves repeatedly encoding and decoding the JSON into different formats to transform it.

By instead peeking into the different variant collections to give a "view" into the render node's Objective-C data we can avoid all that encoding and decoding.

Unfortunately the `docc process-archive index` command still needs to use the slow path.  I filed rdar://128050800 to track fixing that as a separate PR.

## Dependencies

None

## Testing

Build documentation for a large mixed Swift and Objective-C project with main and with this branch and measure the total build time. If the project is large enough the difference should be very noticeable. If you want rough estimate numbers to compare you can use `time` for each call, for example:
```sh
time .build/arm64-apple-macosx/release/docc convert /path/to/SomeCatalog.docc  --additional-symbol-graph-dir /path/to/some-symbol-graph-directory 
```

## Checklist

Make sure you check off the following items. If they cannot be completed, provide a reason.

- [x] Added tests
- [x] Ran the `./bin/test` script and it succeeded
- [x] ~Updated documentation if necessary~ Not applicable