@d-ronnqvist 2024-05-14
Oh, the CI fails because we're not allowed to nest protocols like that yet. I'll move them out to be top-level tomorrow.